chore(Sidebar): move to lab part
