+++ src/components/index.ts
@@ -29,9 +29,9 @@ export { default as Paper } from './Paper'
 export { default as Radio } from './Radio'
 export { default as Select } from './Select'
 export { default as ShowMore } from './ShowMore'
-export { default as Sidebar } from './Sidebar'
-export { default as SidebarItem } from './SidebarItem'
-export { default as SidebarMenu } from './SidebarMenu'
+export { default as Sidebar } from './lab/Sidebar'
+export { default as SidebarItem } from './lab/SidebarItem'
+export { default as SidebarMenu } from './lab/SidebarMenu'
 export { default as Stepper } from './Stepper'
 export { default as Table } from './Table'
 export { default as TableFooter } from './TableFooter'
