+++ src/components/lab/Sidebar/Sidebar.tsx
@@ -2,8 +2,8 @@ import React, { FunctionComponent } from 'react'
 import { withStyles } from '@material-ui/core/styles'
 import cx from 'classnames'
 
-import { StandardProps, PicassoComponent } from '../Picasso'
-import Container from '../Container'
+import { StandardProps, PicassoComponent } from '../../Picasso'
+import Container from '../../Container'
 import SidebarMenu from '../SidebarMenu'
 import SidebarItem from '../SidebarItem'
 import styles from './styles'
