+++ src/components/lab/Sidebar/__image_snapshots__/sidebar-default-snap.png
