+++ src/components/lab/Sidebar/__image_snapshots__/sidebar-links-snap.png
