+++ src/components/lab/Sidebar/story/index.jsx
@@ -1,14 +1,14 @@
 import PicassoBook from '~/.storybook/components/PicassoBook'
 
-import sidebarMenuStory from '@components/SidebarMenu/story'
-import sidebarItemStory from '@components/SidebarItem/story'
+import sidebarMenuStory from '@components/lab/SidebarMenu/story'
+import sidebarItemStory from '@components/lab/SidebarItem/story'
 
 import { Sidebar } from '../Sidebar'
 
 const page = PicassoBook.createPage(
   'Sidebar',
   `Navigation items provide access to parts in your app`,
-  'Layout'
+  'Lab'
 )
 
 page
@@ -19,5 +19,5 @@ page
 
 page
   .createChapter()
-  .addExample('Sidebar/story/Default.example.jsx', 'Default')
-  .addExample('Sidebar/story/Links.example.jsx', 'Links')
+  .addExample('lab/Sidebar/story/Default.example.jsx', 'Default')
+  .addExample('lab/Sidebar/story/Links.example.jsx', 'Links')
